refactor: SortType.kt package 위치 변경 #87
